requirements/development.txt
Django==1.5.1
PIL==1.1.7
South==0.7.4
django-autoslug==1.7.1
django_filter==0.6
psycopg2==2.4
pygeocoder==1.1.4
wsgiref==0.1.2
requests==1.2.3
rq==0.3.11
django-rq==0.5.1
django-mail-factory==0.9
factory_boy==2.1.1
mock-django==0.6.5
coverage==3.5.1
django-coverage==1.2.2
-e git+http://github.com/dominno/django-moderation.git#egg=django-moderation-0.3.2
-e git+http://github.com/JordanReiter/django-jsonfield.git@38e517cf09765d48b5ca2928ada468c70f6cf46e#egg=django-jsonfield-dev